no issue #
when 'src' SGLF skips a tilestep but 'add' SGLF doesn't,
varid needed to be reset. Reset was happening but not
in the proper place, resulting in a non-zero varid
for the first tilestep in the tilestep block.

fixed.

Tests haven't been added in the interest of time.
Tests will be added in another branch.

Arvados-DCO-1.1-Signed-off-by: Abram Connelly <aconnelly@veritasgenetics.com>